[trading_client]: change TimeOnTradeOpenMs to be int64

+++ realtime/trading_client/src/type_defs.go
@@ -26,7 +26,7 @@ type Strategy struct {
 	MinimumTimeBetweenTradesMs int     `json:"minimum_time_between_trades_ms"`
 	KlinesLeftTillAutoclose    int     `json:"klines_left_till_autoclose"`
 	MaximumKlinesHoldTime      int     `json:"maximum_klines_hold_time"`
-	TimeOnTradeOpenMs          int     `json:"time_on_trade_open_ms"`
+	TimeOnTradeOpenMs          int64   `json:"time_on_trade_open_ms"`
 	PriceOnTradeOpen           float64 `json:"price_on_trade_open"`
 	AllocatedSizePerc          float64 `json:"allocated_size_perc"`
 	TakeProfitThresholdPerc    float64 `json:"take_profit_threshold_perc"`
